Move retry code into its own crate
