+++ services-core/Cargo.toml
@@ -14,8 +14,8 @@ byteorder = "1.3.4"
 chrono = { version = "0.4.19", default-features = false  }
 contracts = { path = "../contracts" }
 ethcontract = { version = "0.9",  default-features = false }
-gas-estimation = { path = "../gas-estimation" }
 futures = "0.3.8"
+gas-estimation = { path = "../gas-estimation" }
 isahc = { version = "0.9.13", features = ["json"] }
 lazy_static = "1.4.0"
 log = "0.4.11"
@@ -34,6 +34,7 @@ slog-scope = "4.3.0"
 slog-stdlog = "4.1.0"
 slog-term = "2.6.0"
 thiserror = "1.0"
+transaction-retry = { path = "../transaction-retry" }
 typenum = "1.12.0"
 uint = "0.8"
 url = "2.2.0"
